requirements.txt
selenium
PyYAML
requests
flask